Move RCM state to the subscription manager
