+++ tracer/src/Datadog.Trace/RemoteConfigurationManagement/RcmSubscriptionManager.cs
@@ -3,22 +3,45 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
+#nullable enable
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
 using Datadog.Trace.Logging;
+using Datadog.Trace.RemoteConfigurationManagement.Protocol;
+using Datadog.Trace.Util;
+using Datadog.Trace.Vendors.Serilog.Events;
 
 namespace Datadog.Trace.RemoteConfigurationManagement;
 
 internal class RcmSubscriptionManager : IRcmSubscriptionManager
 {
+    private const int RootVersion = 1;
+
     public static readonly IRcmSubscriptionManager Instance = new RcmSubscriptionManager();
     private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor<RcmSubscriptionManager>();
 
     private readonly List<ISubscription> _subscriptions = new();
     private readonly object _syncRoot = new();
+
+    /// <summary>
+    /// Key is the path
+    /// </summary>
+    private readonly Dictionary<string, RemoteConfigurationCache> _appliedConfigurations = new();
+
+    private readonly string _id;
+
+    private string? _backendClientState;
+    private int _targetsVersion;
     private BigInteger _capabilities;
+
+    public RcmSubscriptionManager()
+    {
+        DatadogLogging.GetLoggerFor<RcmSubscriptionManager>().Debug("New RcmSubscriptionManager");
+        _id = Guid.NewGuid().ToString();
+    }
 
     public bool HasAnySubscription => _subscriptions.Count > 0;
 
@@ -121,6 +144,141 @@ internal class RcmSubscriptionManager : IRcmSubscriptionManager
 #endif
 
         return capabilitiesArray;
+    }
+
+    public GetRcmRequest BuildRequest(RcmClientTracer rcmTracer, string? lastPollError)
+    {
+        lock (_syncRoot)
+        {
+            var cachedTargetFiles = new List<RcmCachedTargetFile>();
+            var configStates = new List<RcmConfigState>();
+            var appliedConfigs = _appliedConfigurations.Values;
+
+            foreach (var cache in appliedConfigs)
+            {
+                cachedTargetFiles.Add(new RcmCachedTargetFile(cache.Path.Path, cache.Length, cache.Hashes.Select(kp => new RcmCachedTargetFileHash(kp.Key, kp.Value)).ToList()));
+                configStates.Add(new RcmConfigState(cache.Path.Id, cache.Version, cache.Path.Product, cache.ApplyState, cache.Error));
+            }
+
+            var rcmState = new RcmClientState(RootVersion, _targetsVersion, configStates, lastPollError != null, lastPollError, _backendClientState);
+            var rcmClient = new RcmClient(_id, ProductKeys, rcmTracer, rcmState, GetCapabilities());
+            var rcmRequest = new GetRcmRequest(rcmClient, cachedTargetFiles);
+
+            return rcmRequest;
+        }
+    }
+
+    public void ProcessResponse(GetRcmResponse response)
+    {
+        if (Log.IsEnabled(LogEventLevel.Debug))
+        {
+            var description = response.TargetFiles.Count > 0
+                                  ? "with the following paths: " + string.Join(",", response.TargetFiles.Select(t => t.Path))
+                                  : "that is empty.";
+            Log.Debug("Received Remote Configuration response {ResponseDescription}.", description);
+        }
+
+        lock (_syncRoot)
+        {
+            _targetsVersion = response.Targets.Signed.Version;
+            _backendClientState = response.Targets.Signed.Custom?.OpaqueBackendState;
+
+            var configByProducts = new Dictionary<string, List<RemoteConfiguration>>();
+            var receivedPaths = new List<string>();
+
+            // handle new configurations
+            foreach (var clientConfig in response.ClientConfigs)
+            {
+                var remoteConfigurationPath = RemoteConfigurationPath.FromPath(clientConfig);
+                receivedPaths.Add(remoteConfigurationPath.Path);
+                var signed = response.Targets.Signed.Targets;
+                var targetFiles =
+                    (response.TargetFiles ?? Enumerable.Empty<RcmFile>()).ToDictionary(f => f.Path, f => f);
+
+                if (!signed.TryGetValue(remoteConfigurationPath.Path, out var signedTarget))
+                {
+                    ThrowHelper.ThrowException($"Missing config {remoteConfigurationPath.Path} in targets");
+                }
+
+                if (!ProductKeys.Contains(remoteConfigurationPath.Product))
+                {
+                    Log.Warning("Received config {RemoteConfigurationPath} for a product that was not requested", remoteConfigurationPath);
+                    continue;
+                }
+
+                var isConfigApplied =
+                    _appliedConfigurations.TryGetValue(remoteConfigurationPath.Path, out var appliedConfig) &&
+                    appliedConfig.Hashes.SequenceEqual(signedTarget.Hashes);
+                if (isConfigApplied)
+                {
+                    continue;
+                }
+
+                if (!targetFiles.TryGetValue(remoteConfigurationPath.Path, out var targetFile))
+                {
+                    ThrowHelper.ThrowException($"Missing config {remoteConfigurationPath.Path} in target files");
+                }
+
+                var remoteConfigurationCache = new RemoteConfigurationCache(remoteConfigurationPath, signedTarget.Length, signedTarget.Hashes, signedTarget.Custom.V);
+                _appliedConfigurations[remoteConfigurationCache.Path.Path] = remoteConfigurationCache;
+
+                var remoteConfiguration = new RemoteConfiguration(remoteConfigurationPath, targetFile.Raw, signedTarget.Length, signedTarget.Hashes, signedTarget.Custom.V);
+                if (!configByProducts.ContainsKey(remoteConfigurationPath.Product))
+                {
+                    configByProducts[remoteConfigurationPath.Product] = new List<RemoteConfiguration>();
+                }
+
+                configByProducts[remoteConfigurationPath.Product].Add(remoteConfiguration);
+            }
+
+            Dictionary<string, List<RemoteConfigurationPath>> removedConfigsByProduct = new();
+            // handle removed configurations
+            foreach (var appliedConfiguration in _appliedConfigurations)
+            {
+                if (receivedPaths.Contains(appliedConfiguration.Key))
+                {
+                    continue;
+                }
+
+                if (!removedConfigsByProduct.ContainsKey(appliedConfiguration.Value.Path.Product))
+                {
+                    removedConfigsByProduct[appliedConfiguration.Value.Path.Product] =
+                        new List<RemoteConfigurationPath>();
+                }
+
+                removedConfigsByProduct[appliedConfiguration.Value.Path.Product].Add(appliedConfiguration.Value.Path);
+            }
+
+            // update applied configurations after removal
+            foreach (var removedConfig in removedConfigsByProduct.Values)
+            {
+                foreach (var value in removedConfig)
+                {
+                    _appliedConfigurations.Remove(value.Path);
+                }
+            }
+
+            var results = Update(configByProducts, removedConfigsByProduct);
+
+            foreach (var result in results)
+            {
+                switch (result.ApplyState)
+                {
+                    case ApplyStates.UNACKNOWLEDGED:
+                        // Do nothing
+                        break;
+                    case ApplyStates.ACKNOWLEDGED:
+                        _appliedConfigurations[result.Filename].Applied();
+                        break;
+                    case ApplyStates.ERROR:
+                        _appliedConfigurations[result.Filename].ErrorOccured(result.Error);
+                        break;
+                    default:
+                        Log.Warning("Unexpected ApplyState: {ApplyState}", result.ApplyState);
+                        break;
+                }
+            }
+        }
     }
 
     private void RefreshProductKeys() => ProductKeys = _subscriptions.SelectMany(s => s.ProductKeys).Distinct().ToList();
